cloud_storage_clients: auth refresh probe per upstream
